honeycomb_render

Module resources

resources used to build the default [App]

Structs

DartGizmos

Dart gizmos configuration group.

DartHeadMul

Rendering option as a resource.

DartRenderColor

Rendering option as a resource.

DartShrink

Rendering option as a resource.

DartWidth

Rendering option as a resource.

EdgeGizmos

Edge gizmos configuration group.

EdgeRenderColor

Rendering option as a resource.

EdgeWidth

Rendering option as a resource.

FaceGizmos

Face gizmos configuration group.

FaceNormals

Collection of normals, organized per faces of a map.

FaceRenderColor

Rendering option as a resource.

FaceShrink

Rendering option as a resource.

Map

Combinatorial map to render.

Map3

Combinatorial map to render.

MapVertices

Collection of a map’s vertices.

VertexGizmos

Vertex gizmos configuration group.

VertexRenderColor

Rendering option as a resource.

VertexWidth

Rendering option as a resource.

VolumeNormals

Collection of normals, organized per volumes of a map.

VolumeRenderColor

Rendering option as a resource.

VolumeShrink

Rendering option as a resource.

WindowVisible

Configuration window status resource.