Crate honeycomb_render

honeycomb-render

This crate implements a graphical debugging tool using the bevy crate. The [App] structure can be used to render a given combinatorial map, using underlying ECS logic to render map items.

All the ECS code used to render maps is left public, allowing advanced user to customize the rendering tool to account for user-defined attributes.

Note that rendering large maps may require running the program in release mode.

Modules

bundles

bundles used to build the default [App]

components

components used to build the default [App]

plugins

plugins used to build the default [App]

resources

resources used to build the default [App]

systems

systems used to build the default [App]

Functions

render_2d_map

Main rendering function.

render_3d_map

Main rendering function.