honeycomb_render
Module systems

systems used to build the default App

Functions

cursor_in_render

Detects if the cursor is positioned in the render tab.

dart_bodies_handles

Dart body mesh handles update system.

dart_bodies_transform

Dart body transform update system.

dart_heads_handle

Dart head mesh handle update system.

dart_heads_transform

Dart head transform update system.

dart_mat_handle

Dart material handle update system.

dart_render

Dart render color and status update system.

draw_inspected_data

Inspection panel drawing function.

draw_options

rendering options drawing function.

edges_handle

Edge mesh handles update system.

edges_mat_handle

Edge material handle update system.

edges_render

Edge render color and status update system.

populate_beta

System used to generate beta entities in the ECS.

populate_darts

System used to generate dart entities in the ECS.

populate_edges

System used to generate edge entities in the ECS.

populate_faces

System used to generate face entities in the ECS.

populate_vertices

System used to generate vertex entities in the ECS.

setup_scene

Scene setup routine.

update_camera

Camera update routine.

update_picking

Picking update routine.

vertices_handle

Vertex mesh handle update system.

vertices_mat_handle

Vertex material handle update system.

vertices_render

Vertex render color and status update system.